refactor: checkIsNotTarget
